fix: return int8

+++ ValueTypeIntegerRange.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity >=0.7.0 <0.9.0;
 contract ValueTypeIntegerRange {
-  function range() public pure returns (int, int) {
-    return (type(int).min, type(int).max);
+  function range() public pure returns (int8, int8) {
+    return (type(int8).min, type(int8).max);
   }
 }
